+++ tests/bootstrap.php
@@ -13,6 +13,7 @@ require KIRJURI_ROOT . '/lib/migrations.php';
 require KIRJURI_ROOT . '/lib/cases.php';
 require KIRJURI_ROOT . '/lib/statistics.php';
 require KIRJURI_ROOT . '/lib/install.php';
+require KIRJURI_ROOT . '/lib/tools.php';
 
 spl_autoload_register(function ($class) {
     $prefix = 'Kirjuri\\Tests\\';